+++ src/main/java/com/cotato/kampus/domain/admin/application/AdminService.java
@@ -16,13 +16,13 @@ import com.cotato.kampus.domain.admin.dto.VerificationPhotoDto;
 import com.cotato.kampus.domain.admin.dto.VerificationWithPhoto;
 import com.cotato.kampus.domain.admin.dto.response.AdminCardNewsPreview;
 import com.cotato.kampus.domain.admin.dto.response.BoardInfo;
-import com.cotato.kampus.domain.board.application.BoardAppender;
-import com.cotato.kampus.domain.board.application.BoardDtoEnhancer;
-import com.cotato.kampus.domain.board.application.BoardFinder;
-import com.cotato.kampus.domain.board.application.BoardUpdater;
-import com.cotato.kampus.domain.board.application.BoardValidator;
-import com.cotato.kampus.domain.board.application.BoardCategoryAppender;
-import com.cotato.kampus.domain.board.dto.BoardDto;
+import com.cotato.kampus.domain.board.implement.BoardAppender;
+import com.cotato.kampus.domain.board.implement.BoardDtoEnhancer;
+import com.cotato.kampus.domain.board.implement.BoardFinder;
+import com.cotato.kampus.domain.board.implement.BoardUpdater;
+import com.cotato.kampus.domain.board.implement.BoardValidator;
+import com.cotato.kampus.domain.board.implement.BoardCategoryAppender;
+import com.cotato.kampus.domain.board.domain.BoardDto;
 import com.cotato.kampus.domain.board.enums.BoardStatus;
 import com.cotato.kampus.domain.board.enums.BoardType;
 import com.cotato.kampus.domain.common.application.ApiUserResolver;
