+++ src/main/java/com/cotato/kampus/domain/admin/dto/response/BoardInfo.java
@@ -1,6 +1,6 @@
 package com.cotato.kampus.domain.admin.dto.response;
 
-import com.cotato.kampus.domain.board.dto.BoardDto;
+import com.cotato.kampus.domain.board.domain.BoardDto;
 import com.cotato.kampus.domain.board.enums.BoardStatus;
 import com.cotato.kampus.domain.board.enums.BoardType;
