+++ src/main/java/com/cotato/kampus/domain/board/api/response/BoardWithDescriptionResponse.java
@@ -1,6 +1,6 @@
-package com.cotato.kampus.domain.board.dto.response;
+package com.cotato.kampus.domain.board.api.response;
 
-import com.cotato.kampus.domain.board.dto.BoardWithFavoriteStatus;
+import com.cotato.kampus.domain.board.domain.BoardWithFavoriteStatus;
 
 public record BoardWithDescriptionResponse(
 	Long boardId,
