+++ src/main/java/com/cotato/kampus/domain/board/dao/repository/BoardCategoryRepository.java
@@ -1,4 +1,4 @@
-package com.cotato.kampus.domain.board.dao;
+package com.cotato.kampus.domain.board.dao.repository;
 
 import java.util.List;
 import java.util.Optional;
@@ -6,7 +6,7 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.cotato.kampus.domain.board.domain.BoardCategory;
+import com.cotato.kampus.domain.board.dao.entity.BoardCategory;
 
 @Repository
 public interface BoardCategoryRepository extends JpaRepository<BoardCategory, Long> {
