+++ src/main/java/com/cotato/kampus/domain/board/implement/BoardCategoryAppender.java
@@ -1,12 +1,12 @@
-package com.cotato.kampus.domain.board.application;
+package com.cotato.kampus.domain.board.implement;
 
 import java.util.List;
 
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.cotato.kampus.domain.board.dao.BoardCategoryRepository;
-import com.cotato.kampus.domain.board.domain.BoardCategory;
+import com.cotato.kampus.domain.board.dao.repository.BoardCategoryRepository;
+import com.cotato.kampus.domain.board.dao.entity.BoardCategory;
 
 import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
