+++ src/main/java/com/cotato/kampus/domain/post/application/PostFinder.java
@@ -9,9 +9,9 @@ import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.cotato.kampus.domain.board.application.BoardFinder;
-import com.cotato.kampus.domain.board.application.BoardCategoryFinder;
-import com.cotato.kampus.domain.board.dto.BoardDto;
+import com.cotato.kampus.domain.board.implement.BoardFinder;
+import com.cotato.kampus.domain.board.implement.BoardCategoryFinder;
+import com.cotato.kampus.domain.board.domain.BoardDto;
 import com.cotato.kampus.domain.common.application.ApiUserResolver;
 import com.cotato.kampus.domain.post.dao.PostCategoryRepository;
 import com.cotato.kampus.domain.post.dao.PostDraftPhotoRepository;
