+++ src/main/java/com/cotato/kampus/domain/post/application/PostValidator.java
@@ -3,12 +3,11 @@ package com.cotato.kampus.domain.post.application;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.cotato.kampus.domain.board.application.BoardFinder;
-import com.cotato.kampus.domain.board.dto.BoardDto;
+import com.cotato.kampus.domain.board.implement.BoardFinder;
+import com.cotato.kampus.domain.board.domain.BoardDto;
 import com.cotato.kampus.domain.board.enums.BoardType;
 import com.cotato.kampus.domain.post.dao.PostScrapRepository;
 import com.cotato.kampus.domain.post.domain.Post;
-import com.cotato.kampus.domain.post.domain.PostDraft;
 import com.cotato.kampus.domain.post.dto.PostDraftDto;
 import com.cotato.kampus.domain.post.dto.PostDto;
 import com.cotato.kampus.global.error.ErrorCode;
